requirements/base.txt
# requirements/base.txt
# Project Chimera - Base Requirements
# Generated from pyproject.toml with: uv pip compile pyproject.toml -o requirements/base.txt

# Core
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
starlette>=0.27.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Async
anyio>=3.7.0
asyncio>=3.4.3

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# ML/AI
openai>=1.3.0
anthropic>=0.7.0
langchain>=0.0.340
langsmith>=0.0.70

# MCP
mcp>=0.1.0

# Utilities
pyyaml>=6.0
jsonschema>=4.19.0
python-dateutil>=2.8.0
pytz>=2023.3

# NOTE: This file is auto-generated. To update:
# 1. Update pyproject.toml
# 2. Run: uv pip compile pyproject.toml -o requirements/base.txt
# 3. Run: uv sync